app: build on linux
