+++ shell.nix
@@ -6,6 +6,10 @@ mkShell {
     gcc
     pkg-config
     udev
+
+    # ggoled_app
+    gtk3.dev
+    # TODO: further deps
   ];
   shellHook = ''
     rustup default stable
